Added migration_role as node that needs database.yml
The node that runs the migrations might not be an app node. For example
a dedicated db node that runs its own migrations needs the Ruby code
and database.yml. In migrations.rake of capistrano-rails, the migrations
can be run on a role specified by the migration_role. This edit is to
ensure that if the migration_role is different from app, that node run
the migrations and connect to the db using database.yml settings.