Changed to release_roles :all as roles that needs database.yml
Other roles besides :app need to be able to access the database. For
example, a dedicated db node that runs its own migrations but is not have
the :app role needs the Ruby code and database.yml. In migrations.rake of
capistrano-rails, the migrations can be run on a role specified by the
migration_role.

This commit is to ensure that all release_roles have a copy of database.yml
on setup, in order that they can connect to the database.